Fix #92 Command fails if a table doesn't have AI PK column.
Blueprint sets up a new Fluent without ensuring it has an array of `columns`,
which is a required attribute. This modifies a call to a `Fluent` constructor
providing a null array `columns.`

Factory declares `$schema` property which never instantiated, and is required to
be an array. This modifies the declaration to provide instantiation of a null array.